requirements.txt
annotated-types==0.7.0
attrs==24.2.0
black==24.8.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
frictionless==4.40.8
greenlet==3.1.1
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
isodate==0.7.2
jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
marko==2.1.2
MarkupSafe==2.1.5
mdurl==0.1.2
multimethod==1.10
mypy-extensions==1.0.0
numpy==1.24.4
packaging==24.1
pandas==2.0.3
pandera==0.20.4
pathspec==0.12.1
petl==1.7.15
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
psycopg2==2.9.9
pydantic==2.9.2
pydantic-core==2.23.4
pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rfc3986==2.0.0
rich==13.9.2
rpds-py==0.20.0
shellingham==1.5.4
simpleeval==0.9.13
six==1.16.0
SQLAlchemy==2.0.35
stringcase==1.2.0
tabulate==0.9.0
text-unidecode==1.3
tomli==2.0.2
typeguard==4.3.0
typer==0.12.5
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
urllib3==2.2.3
validators==0.34.0
wrapt==1.16.0
zipp==3.20.2
